fixed security hole with CSRF and fixed bug with JWT_TOKEN_VALIDITY

+++ CashOverflow/src/main/java/com/revature/util/JwtUtil.java
@@ -26,7 +26,7 @@ public class JwtUtil implements Serializable {
 
 	private static final long serialVersionUID = -1916638954380206609L;
 
-	public static final long JWT_TOKEN_VALIDITY = 5 * 60 * 60;
+	public static final int JWT_TOKEN_VALIDITY = 5 * 60 * 60;
 
 	@Value("${jwt.secret}")
 	private String secret;
